using shorts instead of bytes to get rid of the signed bit

+++ src/main/java/Bus.java
@@ -2,15 +2,15 @@ import exception.MemoryAddressExceedsMemoryException;
 
 public class Bus {
 
-    private byte[] ram;
+    private short[] ram;
     private long sizeInBytes;
 
     public Bus(int sizeInBytes) {
         this.sizeInBytes = sizeInBytes;
-        ram = new byte[sizeInBytes];
+        ram = new short[sizeInBytes];
     }
 
-    public void write(int addr, byte data) {
+    public void write(int addr, short data) {
         if (addr >= 0x0000 && addr <= sizeInBytes) {
             ram[addr] = data;
         } else {
@@ -18,7 +18,7 @@ public class Bus {
         }
     }
 
-    public byte read(int addr, boolean readOnly) {
+    public short read(int addr, boolean readOnly) {
         if (addr >= 0x0000 && addr <= sizeInBytes) {
             return ram[addr];
         } else {
