+++ src/main/java/addressmode/Zp0.java
@@ -3,7 +3,7 @@ package addressmode;
 public class Zp0 extends AddressingMode {
 
     @Override
-    public byte set() {
+    public short set() {
         return 0;
     }
 }
