+++ src/main/java/opcodes/Bvs.java
@@ -5,7 +5,7 @@ package opcodes;
  */
 public class Bvs extends Opcode {
     @Override
-    public byte operate() {
+    public short operate() {
         return 0;
     }
 }
